moved strain cropping transform into waveform_transforms.py, adapted to MFD
