--- dingo/gw/transforms/strain_transforms.py
@@ -1,221 +1,0 @@
-import numpy as np
-from typing import Optional, Union
-
-from dingo.gw.domains import FrequencyDomain
-
-
-class CropMaskStrainRandom(object):
-    """Apply random cropping of strain, by masking waveform and ASD outside the crop."""
-
-    def __init__(
-        self,
-        domain: Union[FrequencyDomain],
-        f_min_upper: Optional[float] = None,
-        f_max_lower: Optional[float] = None,
-        deterministic: bool = False,
-        cropping_probability: float = 1.0,
-        independent_detectors: bool = True,
-        independent_lower_upper: bool = True,
-    ):
-        """
-        Parameters
-        ----------
-        domain: Union[FrequencyDomain]
-            Domain of the waveform data, has to be a frequency domain type.
-        f_min_upper: float
-            New f_min is sampled in range [domain.f_min, f_min_upper].
-            Sampling of f_min is uniform in bins (not in frequency) when the frequency
-            domain is not uniform (e.g., MultibandedFrequencyDomain).
-        f_max_lower: float
-            New f_max is sampled in range [domain.f_max, f_max_lower].
-            Sampling of f_max is uniform in bins (not in frequency) when the frequency
-            domain is not uniform (e.g., MultibandedFrequencyDomain).
-        deterministic: bool
-            If True, don't sample truncation range, but instead always truncate to range
-            [f_min_upper, f_max_lower]. This is used for inference.
-        cropping_probability: float
-            probability for a given sample to be cropped
-        independent_detectors: bool
-            If True, crop boundaries are sampled independently for different detectors.
-        independent_lower_upper: bool
-            If True, the cropping probability is applied to lower and upper boundaries
-            individually. If False, then with a probability of P = cropping_probability
-            both lower and upper cropping is applied, and with 1-P, no cropping is
-            applied from either direction.
-        """
-        self.check_inputs(
-            domain, f_min_upper, f_max_lower, cropping_probability, deterministic
-        )
-        self._deterministic = deterministic
-        self.cropping_probability = cropping_probability
-        self.independent_detectors = independent_detectors
-        self.independent_lower_upper = independent_lower_upper
-        frequencies = domain()[domain.min_idx :]
-        self.len_domain = len(frequencies)
-
-        if f_max_lower is not None:
-            self._idx_bound_f_max = np.argmin(np.abs(f_max_lower - frequencies))
-        else:
-            self._idx_bound_f_max = self.len_domain - 1
-
-        if f_min_upper is not None:
-            self._idx_bound_f_min = np.argmin(np.abs(f_min_upper - frequencies))
-        else:
-            self._idx_bound_f_min = 0
-
-    def sample_upper_bound_indices(self, shape):
-        """Sample indices for upper crop boundaries."""
-        if self._deterministic:
-            return np.ones(shape) * self._idx_bound_f_max
-        else:
-            return np.random.randint(self._idx_bound_f_max, self.len_domain, shape)
-
-    def sample_lower_bound_indices(self, shape):
-        """Sample indices for lower crop boundaries."""
-        if self._deterministic:
-            return np.ones(shape) * self._idx_bound_f_min
-        else:
-            # self._idx_bound_f_min is inclusive bound, so need to add 1
-            return np.random.randint(0, self._idx_bound_f_min + 1, shape)
-
-    def check_inputs(
-        self, domain, f_min_upper, f_max_lower, cropping_probability, deterministic
-    ):
-        # check domain
-        if not isinstance(domain, (FrequencyDomain,)):
-            raise ValueError(
-                f"Domain should be a frequency domain type, got {type(domain)}."
-            )
-        # check validity of ranges
-        if f_min_upper is not None:
-            if not domain.f_min < f_min_upper < domain.f_max:
-                raise ValueError(
-                    f"Expected f_min_upper in domain range [{domain.f_min},"
-                    f" {domain.f_max}], got {f_min_upper}."
-                )
-        if f_max_lower is not None:
-            if not domain.f_min < f_max_lower < domain.f_max:
-                raise ValueError(
-                    f"Expected f_max_lower in domain range [{domain.f_min},"
-                    f" {domain.f_max}], got {f_max_lower}."
-                )
-        if f_min_upper and f_max_lower and f_min_upper >= f_max_lower:
-            raise ValueError(
-                f"Expected f_min_upper < f_max_lower, got {f_min_upper}, {f_max_lower}."
-            )
-        if not 0 <= cropping_probability <= 1.0:
-            raise ValueError(
-                f"Cropping probability should be in [0, 1], got {cropping_probability}."
-            )
-        # check that no non-trivial cropping probability is set when deterministic = True
-        if deterministic and cropping_probability < 1.0:
-            raise ValueError(
-                f"cropping_probability must be 1.0 when deterministic = True, got "
-                f"{cropping_probability}."
-            )
-
-    def __call__(self, input_sample: dict):
-        """
-        Parameters
-        ----------
-        input_sample : dict
-            sample["waveform"]: Dict with values being arrays containing waveforms,
-            or torch Tensor with the waveform.
-
-        Returns
-        -------
-        dict of the same form as the input, but with transformed (crop-masked) waveforms.
-        """
-        sample = input_sample.copy()
-        strain = sample["waveform"]
-        if strain.shape[-1] != self.len_domain:
-            raise ValueError(
-                f"Expected waveform input of shape [..., {self.len_domain}], "
-                f"got {strain.shape}."
-            )
-
-        # Sample boundary indices for crops
-        constant_ax = 3 - self.independent_detectors
-        lower = self.sample_lower_bound_indices(strain.shape[:-constant_ax])
-        upper = self.sample_upper_bound_indices(strain.shape[:-constant_ax])
-
-        # Only apply crops to a fraction of self.cropping_probability
-        if self.cropping_probability < 1:
-            mask = np.random.uniform(size=lower.shape) <= self.cropping_probability
-            lower = np.where(mask, lower, 0)
-            if self.independent_lower_upper:
-                mask = np.random.uniform(size=lower.shape) <= self.cropping_probability
-            upper = np.where(mask, upper, self.len_domain)
-
-        # Broadcast boundaries and apply cropping
-        mask_lower = np.arange(self.len_domain) >= lower[(...,) + (None,) * constant_ax]
-        mask_upper = np.arange(self.len_domain) <= upper[(...,) + (None,) * constant_ax]
-        strain = np.where(mask_lower, strain, 0)
-        strain = np.where(mask_upper, strain, 0)
-        sample["waveform"] = strain
-
-        return sample
-
-
-class CropMaskStrain(object):
-    """
-    Crop or mask strain with zeros depending on minimum_frequency and maximum_frequency.
-    """
-
-    def __init__(
-        self,
-        minimum_frequency: Optional[float | dict[str, float]] = None,
-        maximum_frequency: Optional[float | dict[str, float]] = None,
-        print_output: bool = False,
-    ):
-        """
-        Parameters
-        ----------
-        minimum_frequency: Optional[float | dict[str, float]]
-            Update of f_min, if float, the same value will be used for all detectors.
-        maximum_frequency: Optional[float | dict[str, float]]
-            Update of f_max, if float, the same value will be used for all detectors.
-        print_output: bool
-            Whether to write print statements to the console.
-        """
-        self.minimum_frequency = minimum_frequency
-        self.maximum_frequency = maximum_frequency
-
-        if print_output:
-            print(
-                f"Transform UpdateFrequencyRange activated:"
-                f"  Settings: \n"
-                f"    - Minimum_frequency update: {self.minimum_frequency}\n"
-                f"    - Maximum_frequency update: {self.maximum_frequency}\n"
-            )
-
-    def __call__(self, input_sample):
-        """
-        Parameters
-        ----------
-        input_sample: Dict
-            Values for keys
-            - 'waveform':
-            Sample of shape [batch_size, num_tokens, num_features]
-            - 'position', shape [batch_size, num_tokens, 3]
-               contains information [f_min, f_max, block]
-            - 'drop_token_mask', shape [batch_size, num_tokens]
-
-        Returns
-        ----------
-        sample: Dict
-            input_sample with modified value for key
-            - 'drop_token_mask', shape [batch_size, num_tokens]
-
-        """
-        sample = input_sample.copy()
-
-        # Update minimum_frequency
-        if self.minimum_frequency is not None:
-            raise ValueError("Update minimum_frequency is not yet implemented.")
-
-        # Update maximum_frequency
-        if self.maximum_frequency is not None:
-            raise ValueError("Update maximum_frequency is not yet implemented.")
-
-        return sample
